Moving object printer to main.
